requirements.txt
ipfshttpclient
web3
streamlit
requests
python-dotenv
fastapi